refactor: move e2e test files where protractor expects to find them
