+++ e2e/_unused/broken-example.html
@@ -57,7 +57,7 @@
         </li>
       </ul>
     </div>
-    <script src="../node_modules/angular/angular.js"></script>
+    <script src="../../node_modules/angular/angular.js"></script>
     <script>
       angular.module('bpAppModule', []);
         /**
