+++ e2e/_unused/correct-example.html
@@ -57,8 +57,8 @@ correctly uses only one module with `ng-app`. -->
     </ul>
   </div>
 
-  <script src="../node_modules/angular/angular.js"></script>
-  <script src="../dist/hint.js"></script>
+  <script src="../../node_modules/angular/angular.js"></script>
+  <script src="../../dist/hint.js"></script>
   <script>
     angular.module('bpAppModule', [])
       /**
