+++ e2e/_unused/manual-bootstrap.html
@@ -5,7 +5,7 @@
     <title>Angular Hint Example</title>
 
     <script src="../../node_modules/angular/angular.js"></script>
-    <script src="../../bower_components/angular-route/angular-route.js"></script>
+    <script src="../../node_modules/angular-route/angular-route.js"></script>
     <script src="../../dist/hint.js"></script>
     <script src="../util.js"></script>
