+++ e2e/_unused/profile.html
@@ -34,8 +34,8 @@
     </ul>
   </div>
 
-  <script src="../node_modules/angular/angular.js"></script>
-  <script src="../dist/hint.js"></script>
+  <script src="../../node_modules/angular/angular.js"></script>
+  <script src="../../dist/hint.js"></script>
   <script>
     angular.module('profileApp', [])
       /**
