+++ e2e/exclude-wrong-module-name/exclude-wrong-module-name.spec.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var consoleText = require('./util.protractor');
+var consoleText = require('../util.protractor');
 
 describe('angularHint', function() {
   it('should warn if ng-hint is called with unknown options', function() {
