+++ e2e/no-hint/index.html
@@ -1,0 +1,14 @@
+<!doctype html>
+<html ng-app>
+<head>
+  <meta charset="utf-8">
+  <title>Angular Hint Example</title>
+</head>
+<body>
+  <div id="console"></div>
+
+  <script src="../../node_modules/angular/angular.js"></script>
+  <script src="../../dist/hint.js"></script>
+  <script src="../util.js"></script>
+</body>
+</html>
